requirements.txt
beautifulsoup4==4.12.3
lxml==5.3.0
pytest-playwright==0.6.2
python-dotenv==1.0.1
pathvalidate==3.2.1